requirements.txt
numpy ==1.24.*
pandas ==1.5.*
pytest ==7.3.*
pyyaml ==6.0.*
scipy ==1.10.*
setuptools ==67.6.*
wheel ==0.40.*
plink
